updated workspace credits struct to i64, made payment_method as optional string, multiplied credits by 10

+++ api/src/routes/api.patr.cloud/workspace/billing.rs
@@ -830,7 +830,7 @@ async fn get_credits(
 	.into_iter()
 	.map(|credits| WorkspaceCredits {
 		workspace_id: credits.workspace_id,
-		credits: credits.credits,
+		credits: credits.credits as f64,
 		metadata: credits.metadata,
 		date: credits.date,
 	})
